refactor(testing): split ast_assertions.rs into modular per-element structure
Split the 1,344-line ast_assertions.rs into a modular directory structure
with dedicated modules for each AST element type:

- ast_assertions/mod.rs: Entry point + ContentItemAssertion (259 lines)
- ast_assertions/assertions/: Element-specific assertion modules
  - mod.rs: Re-exports + helper functions
  - document.rs: DocumentAssertion
  - paragraph.rs: ParagraphAssertion
  - session.rs: SessionAssertion
  - list.rs: ListAssertion + ListItemAssertion
  - definition.rs: DefinitionAssertion
  - annotation.rs: AnnotationAssertion (with tests)
  - verbatim.rs: VerbatimBlockAssertion + VerbatimGroupAssertion
  - children.rs: ChildrenAssertion

Benefits:
- Improved code navigation and maintainability
- Each element type in its own focused module
- Easier to modify individual assertion types independently
- Tests kept with their respective assertion modules
- Main module reduced from 1,344 to 259 lines
- Consistent with modular patterns used elsewhere in the codebase

All tests passing (253 lib tests + 13 doc tests + 75 integration tests).
